requirements.txt
selenium
dotenv
tqdm
time
telebot
pillow
setuptools